moved hooks into utils
